Fix URLRecord not defining all fields on initialisation
When URLRecord.__init__ is executed, it first calls super().__init__, which resolves to URLPriorities's init method. However, because URLPriorities.__init__ does not call super().__init__, it doesn't go further in the hierarchy, therefore never defining the fields of URLData and URLResult.

NB: There's still a bug in that URLRecord.database_attributes is incorrect, but since that class is never used as a URLDatabaseMixin, that should not be a problem. It might be worth overwriting the database_items method though to prevent such usage explicitly.